django_backend/requirements.txt
# Django Backend Requirements for TIDES Project & Scheme System

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7  # PostgreSQL adapter
# or for MySQL: mysqlclient==2.2.0

# File handling
Pillow==10.0.1
django-storages==1.14.2

# API Documentation
drf-spectacular==0.26.5

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# Report Generation
reportlab==4.0.4
openpyxl==3.1.2

# Date/Time handling
python-dateutil==2.8.2

# Utilities
requests==2.31.0
celery==5.3.4  # For background tasks
redis==5.0.1   # For caching and task queue

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
factory-boy==3.3.0
pytest-django==4.6.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0